snipt/requirements.txt
BeautifulSoup
Django
django-annoying
django-bcrypt
django-css
django-grappelli
django-pagination
django-postmark
django-registration
django-taggit
django-tastypie
django-templatetag-sugar
gunicorn
johnny-cache
lxml
psycopg2
Pygments
python-memcached
pyyaml
South
uuid
Werkzeug
